nike/api/requirements.txt
fastapi
uvicorn
sqlalchemy
python-box
dependency-injector
pydantic-settings
cryptography
PyJWT
mysqlclient==2.1.1
passlib
bcrypt
python-multipart
python-jose
aiofiles
pytest
jinja2
websockets
pytesseract
tesseract
pytz
python-json-logger
simplejson